Fix token amount calculation precision in PayPal order capture
- Replace Number.parseInt with Number.parseFloat for more accurate token amount calculation
- Ensure precise decimal handling when converting fiat amount to token amount

+++ packages/functions/src/api/v1/[paymentId]/paypal-capture-order.ts
@@ -127,8 +127,8 @@ const handleOnChainTransaction = async (
           transaction.finalAmountInFiat - transaction.platformFeesInFiat;
         finalPayoutAmount = finalPayoutAmount - transaction.paymentProcessFeesInFiat
         transaction.finalAmountInFiat = finalPayoutAmount
-        transaction.amountInToken = Number.parseInt(removePercentageFromNumber(
-          parseInt(
+        transaction.amountInToken = Number.parseFloat(removePercentageFromNumber(
+          parseFloat(
             (
               transaction.finalAmountInFiat *
               10 ** transaction.decimals
@@ -162,8 +162,8 @@ const handleOnChainTransaction = async (
           transaction.finalAmountInFiat - transaction.platformFeesInFiat;
         finalPayoutAmount = finalPayoutAmount - transaction.paymentProcessFeesInFiat
         transaction.finalAmountInFiat = finalPayoutAmount
-        transaction.amountInToken = Number.parseInt(removePercentageFromNumber(
-          parseInt(
+        transaction.amountInToken = Number.parseFloat(removePercentageFromNumber(
+          parseFloat(
             (
               transaction.finalAmountInFiat *
               10 ** transaction.decimals
@@ -201,8 +201,8 @@ const handleOnChainTransaction = async (
           transaction.finalAmountInFiat - transaction.platformFeesInFiat;
         finalPayoutAmount = finalPayoutAmount - transaction.paymentProcessFeesInFiat
         transaction.finalAmountInFiat = finalPayoutAmount
-        transaction.amountInToken = Number.parseInt(removePercentageFromNumber(
-          parseInt(
+        transaction.amountInToken = Number.parseFloat(removePercentageFromNumber(
+          parseFloat(
             (
               transaction.finalAmountInFiat *
               10 ** transaction.decimals
@@ -277,8 +277,8 @@ const handleOnChainTransaction = async (
           transaction.finalAmountInFiat,
           project.feePercentage
         )
-        transaction.amountInToken = Number.parseInt(removePercentageFromNumber(
-          parseInt(
+        transaction.amountInToken = Number.parseFloat(removePercentageFromNumber(
+          parseFloat(
             (
               transaction.finalAmountInFiat *
               10 ** transaction.decimals
